Refactor: ImmutableSet to new file
